+++ aqua/core/__init__.py
@@ -3,6 +3,7 @@ from .version import __version__
 from .graphics import plot_single_map, plot_maps, plot_single_map_diff, plot_timeseries
 from .graphics import plot_hovmoller
 from .graphics import plot_lat_lon_profiles, plot_seasonal_lat_lon_profiles
+from .catgen import AquaFDBGenerator
 from .drop import Drop
 from .reader import Reader, Streaming, show_catalog_content
 from .regridder import Regridder
@@ -16,6 +17,7 @@ from .histogram import histogram
 __all__ = ["plot_single_map", "plot_maps", "plot_single_map_diff", "plot_timeseries",
            "plot_hovmoller", "histogram",
            "plot_lat_lon_profiles", "plot_seasonal_lat_lon_profiles",
+           "AquaFDBGenerator",
            "Drop",
            "Reader", "Streaming", "show_catalog_content",
            "Regridder", "GridBuilder",
