+++ aqua/core/catgen/__init__.py
@@ -1,0 +1,4 @@
+"""CatGen module."""
+from .catgen import AquaFDBGenerator, get_nested
+
+__all__ = ["AquaFDBGenerator", "get_nested"]
